requirements.txt
cupy-cuda11x==13.0.0
dask==2023.11.0
h5py==3.8.0
hydra-core==1.3.2
matplotlib==3.8.3
mp-pyrho==0.3.0
omegaconf==2.3.0
pandas==1.5.3
plotly==5.13.0
pybader==0.3.12
pymatgen>=2024.3.1
pytorch-lightning>=2.2.0
pyyaml==6.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
sparse==0.15.1
torch==2.2.0
tqdm==4.64.1
wandb==0.16.3
